+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
-- See git history for changes.
+### Added
+- Use `bedrock-credentials-rest` to display credentials.
 
 [Unreleased]: https://github.com/digitalbazaar/bedrock-credential-curator/compare/0.0.0...HEAD